refactor: narrow projection schema, incremental ingest, evidence-based handoff
Reshape the store around a narrower provider-neutral projection with
incremental ingest and maintenance ops (db checkpoint/optimize/reindex),
provider-scoped skills/MCP metadata reconcile, aligned provider invocation
fields, and a models rollup. Add the evidence-based handoff package for
cross-agent recovery, with inlined output bounded by --max-output-bytes.

Old databases are rebuilt, not migrated in place: the projection epoch is
bumped, so any database whose stamp differs is wiped and rebuilt from the
transcripts (the source of truth) on next open. Dead started_at indexes and
dead columns are dropped in the projection-cleanup migration.

Review hardening:
- retention: drop the reference to the removed turns.summary column so the
  privacy/balanced redaction pass no longer aborts at runtime
- handoff: classify interrupted (non-success) agent runs as incomplete, not
  completed, and emit them as non-authoritative evidence
- codex parser: fold pending token_count tokens into a kept turn that ends
  without an invocation, so they are not dropped from the totals
- httpapi: do not drop a stored session from /v1/status pagination behind a
  live-only overlay; surface the WAL-checkpoint busy flag with a nonzero exit;
  pair the replay-gap id with its reason atomically; drop the one-way persist
  block so durability self-recovers
- collector: reconcile skills and MCP metadata per-kind, so one scan's failure
  no longer suppresses the other's reconcile
- textutil/handoff: rune-safe truncation and clipping; share the truncate,
  first-non-blank, and tool-call clip helpers instead of re-rolling them
- assorted cleanups: idiomatic boolean toggle, dead timeline fields, defer-based
  context cancel, accessor consistency, inlined single-use chunk helper

+++ internal/collector/inode_unix.go
@@ -10,9 +10,9 @@ import (
 // fileInode returns the inode number from a stat result on unix-like systems.
 // Folding it into the change-detection fingerprint lets rotation (a recreated
 // file reusing the same path) be detected even when size and mtime coincide.
-func fileInode(info os.FileInfo) uint64 {
+func fileInode(info os.FileInfo) int64 {
 	if st, ok := info.Sys().(*syscall.Stat_t); ok {
-		return st.Ino
+		return int64(st.Ino)
 	}
 	return 0
 }
